New tool to ingest crashes via the core_pattern
The new 'postmortem' tool for collecting stacktraces from a crashing
program. Derived off of the existing eu-stack tool, it uses parts of the
core file to collect the inital value of the registers; Then uses
vm_readv to walk the stack. This is especially useful when the entire
core file can not fit in-memory nor be written to disk.